@@ -1,8 +1,9 @@
-//+build android,mobilebind
+//go:build android && mobilebind
+// +build android,mobilebind
 
 package engo
 
-//TouchEvent handles the touch events sent from Android and puts them in the InputManager
+// TouchEvent handles the touch events sent from Android and puts them in the InputManager
 func TouchEvent(x, y, id, action int) {
 	Input.Mouse.X = float32(x) / opts.GlobalScale.X
 	Input.Mouse.Y = float32(y) / opts.GlobalScale.Y
