@@ -1,4 +1,5 @@
-//+build js
+//go:build js
+// +build js
 
 package engo
 
@@ -52,8 +53,8 @@ func (gm *GamepadManager) registerGamepadImpl(name string) error {
 
 func (gm *GamepadManager) updateImpl() {
 	if window.IsUndefined() || window.Get("navigator").IsUndefined() {
-               return // node for testing
-        }
+		return // node for testing
+	}
 	gpds := window.Get("navigator").Call("getGamepads")
 	gm.mutex.Lock()
 	defer gm.mutex.Unlock()
